+++ src/aks-agent/HISTORY.rst
@@ -14,7 +14,7 @@ Pending
 
 1.0.0b16
 ++++++++
-* Fix: correct wrong prompt message f0r init and cleanup
+* Fix: correct wrong prompt message for init and cleanup
 * Fix: prompt the whole flags including --resource-group, --name and optional --namespace for az aks agent command
 * Improve: cluster mode cleanup will wait for pods to be removed after deletion